separate `execute` and `subscribe` (#5)
This PR separates the `execute` and `subscribe` functions as well as the `Executor` and `Subscriber` types.

It appears that incremental results returned by `execute` or an `Executor` will not actually match that of `subscribe` or a `Subscriber`.

The stitcher will only require use of an `Executor`, not a `Subscriber`. We can separate the functions and types a this point for simplicity.