deprecate current Stitcher
I'm not happy with the current proto-stitcher which as defined only allows supplementing a single object at a time with fields from different services.

It should be possibly to kick off a worker that can supplement multiple objects at once.

Since the proto-stitcher doesn't do anything yet, we can remove it for now and add a different format later.