use AsyncGenerator signature instead of AsyncIterableIterator
previously, we used AsyncIterableIterator because graphql-js actually returns an object with the AsyncGenerator signature that isn't actually a generator, i.e. does not resolve all calls to `.next()`, `.throw()`, and `.return()` in call order.

Now that this library uses a Repeater-based implementation of mapAsyncIterable, we always actually return a generator, so we can be less strict about this distinction.