add PlanResult abstraction (#53)
Fields for an operation are split to different subschemas, and the results are merged.

Previously, we would wait to merge until the executors for all subschemas returned; now, we merge results as they return, including the asyncIterators. This requires an abstraction for merging asyncIterators dynamically, which is accomplished by the new Consolidator abstraction.

This same abstraction will allow us to merge in stitched asyncIterators as they return just by pushing the mapped asyncIterators to the consolidator.