polish: revamp plan nomenclature (#129)
Naming! It's difficult.

With these changes, hopefully the mental model of our plans is better represented. A RootPlan contains a set of  SubschemaPlans for each individual subschema, which in turn have a FieldTree, a map of fields to type-dependent  SubschemaPlans necessary to patch a subschema response with the data from the other subschemas necessary to complete the request as well as a nested fieldTree to do the same for subfields from the original subschema.

Another advantage is less overloading of the stitch term!